Rename to generalize and move to regular code
