--- Content.Shared/_RMC14/Construction/RMCDeconstructComponent.cs
@@ -1,11 +1,0 @@
-﻿namespace Content.Shared._RMC14.Construction;
-
-[RegisterComponent]
-public sealed partial class RMCDeconstructComponent : Component
-{
-    [DataField]
-    public string? VerbText;
-
-    [DataField]
-    public string? VerbFailedText;
-}
