+++ Resources/Locale/en-US/_RMC14/construction/deconstruct.ftl
@@ -1,4 +1,4 @@
-﻿rmc-deconstruct-verb = Deconstruct
-rmc-deconstruct-verb-failed = Can't deconstruct.
-rmc-deconstruct-verb-fold = Fold box
-rmc-deconstruct-verb-fold-failed = {$entityName} can't be folded.
+﻿construction-interaction-component-verb = Deconstruct
+construction-interaction-component-verb-failed = Can't deconstruct.
+rmc-construction-verb-fold = Fold box
+rmc-construction-verb-fold-failed = {$entityName} can't be folded.
